Basic Auto Leveling (#18)

+++ build.bat
@@ -2,15 +2,16 @@
 
 echo Start building Koolo
 echo Cleaning up previous artifacts...
-rmdir /s /q build > NUL || goto :error
+if exist build rmdir /s /q build > NUL || goto :error
 
 echo Building Koolo binary...
-go build -tags static --ldflags -extldflags="-static" -o build/koolo.exe ./cmd/koolo/main.go > NUL || goto :error
+go build -trimpath -tags static --ldflags -extldflags="-static" -o build/koolo.exe ./cmd/koolo/main.go > NUL || goto :error
 
 echo Copying assets...
 mkdir build\config > NUL || goto :error
 copy config\config.yaml.dist build\config\config.yaml  > NUL || goto :error
 xcopy /q /E /I /y config\pickit build\config\pickit  > NUL || goto :error
+xcopy /q /E /I /y assets build\assets  > NUL || goto :error
 xcopy /q /y rustdecrypt.dll build  > NUL || goto :error
 xcopy /q /y koolo-map.exe build > NUL || goto :error
 xcopy /q /y d2.install.reg build > NUL || goto :error
